requirements.txt
openai>=1.3.0
pyaudio>=0.2.11
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.0
SpeechRecognition>=3.10.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
